Revert "Add common/ directory for architecture-independent tests"
This reverts commit 1f7d5d9617888edceafb60a1f8ba4cd0b809ca89.
